+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+# v2.1.4
+
+### Certificates
+
+#### Fixes
+* b0d1a49 fix(certificates): `keyfactor_certificate` data and resource types will not store auto password used to recover private key. `auto_password` has been removed from schema and state.
+* b0d1a49 fix(certificates): `keyfactor_certificate` resource type will no longer trigger replacement if `key_password` is changed. #74 #79 #80
+* b0d1a49 fix(certificates): When looking up a certificate by CN, `IncludeHasPrivateKey` is now included in the call to the Command API.
+* b0d1a49 fix(certificates): `keyfactor_certificate` resource updates `ca_cert` use correct field.
+* b0d1a49 fix(certificates): `keyfactor_certificate` resource updates `key_password` will now use plan value.
+* b0d1a49 fix(certificates): `keyfactor_certificate` resource updates `certificate_id` field now included using state value.
+* b0d1a49 fix(certificates): When sorting SAN lists, if length varies don't even try to sort as there is obviously a change and replacement must be triggered.
+
+# v2.1.3
+### Certificates
+
+#### Fixes
+* bb5498d fix(certificates): Sort SANs in the same order as state when they come back from the Command API. #66
+
+# v2.1.2
+### Certificates
+
+#### Fixes
+* e0f6c7c fix(certificates): Sort SANs when they come back from the Command API. #66
+
+# v2.1.1
+### Certificates
+
+#### Fixes
+* 0f5d1fe fix(certificates): `key_password` now takes correct precedence #72 #75
+* 594677d fix(certificates): Treat deleted certs as needing replacement. #73
+
 # v2.1.0
 ### Certificates